onetopia/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
SQLAlchemy==2.0.36
psycopg2==2.9.11